+++ src/dark/__init__.py
@@ -1,6 +1,7 @@
 import sys
 from importlib.metadata import PackageNotFoundError, version
-from typing import TextIO
+from pathlib import Path
+from typing import BinaryIO, TextIO
 
 if sys.version_info < (3, 10):
     raise Exception("The dark matter code needs Python 3.10 or later.")  # pyright: ignore[reportUnreachable]
@@ -13,4 +14,4 @@ except PackageNotFoundError:
     __version__ = "unknown"
 
 
-File = TextIO | str
+File = TextIO | BinaryIO | str | bytes | Path